requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Cython==0.29.21
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.4
django-cors-headers==3.7.0
django-import-export==2.5.0
django-simpleui==2021.6.2
et-xmlfile==1.1.0
idna==2.10
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.1.5
psycopg2-binary==2.9.1
py3Fdfs==2.2.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.3
xlrd==1.2.0
xlwt==1.3.0
